builtin cd : les messages d’erreurs sont sur STDERR, ajout de la fonctionnalité ‘cd #’ [5 janv. 2022 à 11:59]
